apps/checkout/requirements.txt
requests
python-dotenv
azure-monitor-opentelemetry
opentelemetry-api
opentelemetry-sdk